Create SeasonBuilder
24 July 2021 - Sample Python that builds a list of all unique seasons.   print(uniqueseasons) returns a list of all possible unique seasons with unique matchups.  There are 105 unique sets of weekly matchups in the 8 team league and from those 105 possible weeks, 416 truly unique Single Round Robin Tournament (SRRT) seasons.  

Running this snippet will take a while.  the incrementing of a+=200 cuts down on the number of loops executed by 99%.  Incrementing a+=1 caused the loop to take 6+ hours to generate 31M possible seasons outstanding.   I have not experimented with 10 teams, as this would cause nesting 9 levels deep!! 

As of 24 July 2021 , I will upload the code that generates the weeks[] list  into a separate file.  So I have included it in here statically.